requirements-dev.txt
# This requirements file should only include dependencies for development
pathlib_mate                            # autopep8 your code
twine                                   # make distribution archive
wheel                                   # make pre-compiled distribution package
build                                   # build distribution package
boto_session_manager>=1.5.3,<2.0.0
s3pathlib>=2.0.1,<3.0.0
# note: don't install pandas, your local pandas may conflict with Glue runtime pandas
pyspark==3.3.0
aws-glue-libs @ git+https://github.com/awslabs/aws-glue-libs.git@v4.0
awswrangler>=3.0.0,<4.0.0
